+++ apps/mobile/src/hooks/sticky-nav.ts
@@ -3,10 +3,13 @@ import { useRef } from 'react'
 import { type NativeScrollEvent, type NativeSyntheticEvent } from 'react-native'
 
 import { mitter } from '~/lib/mitt'
+import { usePreferences } from '~/stores/preferences'
 
 export function useStickyNav() {
+  const { hideHeaderOnScroll, hideTabBarOnScroll } = usePreferences()
+
   const previous = useRef(0)
-  const last = useRef<'show-nav' | 'hide-nav' | undefined>(undefined)
+  const last = useRef<'show' | 'hide' | undefined>(undefined)
 
   function onScroll(event: NativeSyntheticEvent<NativeScrollEvent>) {
     const y = clamp(
@@ -17,14 +20,19 @@ export function useStickyNav() {
 
     const delta = y - previous.current
 
-    const next = delta > 0 ? 'hide-nav' : 'show-nav'
+    const next = delta > 0 ? 'hide' : 'show'
 
     if (next !== last.current) {
-      mitter.emit(next)
+      if (hideHeaderOnScroll) {
+        mitter.emit(next === 'hide' ? 'hide-header' : 'show-header')
+      }
+
+      if (hideTabBarOnScroll) {
+        mitter.emit(next === 'hide' ? 'hide-tab-bar' : 'show-tab-bar')
+      }
     }
 
     last.current = next
-
     previous.current = y
   }